% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitcreds-package.R
\docType{package}
\name{gitcreds-package}
\alias{gitcreds-package}
\alias{_PACKAGE}
\title{gitcreds: Query 'git' Credentials from 'R'}
\description{
Query, set, delete credentials from the 'git' credential store. Manage 'GitHub' tokens and other 'git' credentials. This package is to be used by other packages that need to authenticate to 'GitHub' and/or other 'git' repositories.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitcreds.r-lib.org/}
  \item \url{https://github.com/r-lib/gitcreds}
  \item Report bugs at \url{https://github.com/r-lib/gitcreds/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
