% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-auth.R
\name{gitcreds_get_cache}
\alias{gitcreds_get_cache}
\title{Parse a gitcreds cache environment variable}
\usage{
gitcreds_get_cache(ev)
}
\arguments{
\item{ev}{Environment variable to parse.}
}
\value{
A \code{gitcreds}object, or \code{NULL}, if the environment variable
is not set, or contains an invalid cache record. For an invalid
record it will also throw a warning.
}
\description{
Parse a gitcreds cache environment variable
}
