% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-auth.R
\name{gitcreds_get_cmd}
\alias{gitcreds_get_cmd}
\title{The command that \code{\link[=gitcreds_get]{gitcreds_get()}} would run}
\usage{
gitcreds_get_cmd(url = "https://github.com")
}
\arguments{
\item{url}{URL to get, set or delete credentials for. It may contain a
user name, which is typically (but not always) used by the credential
helpers. It may also contain a path, which is typically (but not always)
ignored by the credential helpers.}
}
\description{
It is sometimes useful to get the system command that \code{\link[=gitcreds_get]{gitcreds_get()}}
would run, to run it in a different way, e.g. in a background process.
}
\details{
Differences with \code{\link[=gitcreds_get]{gitcreds_get()}}:
\itemize{
\item Does not deal with cache environment variables. You can do that
manually with \code{\link[=gitcreds_cache_envvar]{gitcreds_cache_envvar()}} and \code{\link[=gitcreds_get_cache]{gitcreds_get_cache()}}.
\item Does not check for git separately, so you will never get a
\code{gitcreds_nogit_error} error. You can run \code{git --version} if you
want to check for git.
}
}
