% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-auth.R
\name{gitcreds_list_helpers}
\alias{gitcreds_list_helpers}
\title{List active git credential helpers}
\usage{
gitcreds_list_helpers()
}
\value{
Character vector, corresponding to the `credential.helper`
git configuration key. Usually it contains a single credential helper,
but it is possible to configure multiple helpers.
}
\description{
List active git credential helpers
}
