% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-doc.R
\name{gitcreds_parse_output}
\alias{gitcreds_parse_output}
\title{Parse standard output from \verb{git credential fill}}
\usage{
gitcreds_parse_output(txt, url)
}
\arguments{
\item{txt}{Character vector, standard output lines from
\verb{git credential fill}.}

\item{url}{URL we queried, to be able to create a better error message.}
}
\value{
\code{gitcreds} object.
}
\description{
Parse standard output from \verb{git credential fill}
}
\details{
For dummy credentials (i.e. the lack of credentials), it throws an
error of class \code{gitcreds_no_credentials}.
}
