% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-auth.R
\name{gitcreds_cache_envvar}
\alias{gitcreds_cache_envvar}
\title{Environment variable to cache the password for a URL}
\usage{
gitcreds_cache_envvar(url)
}
\arguments{
\item{url}{Character vector of URLs, they may contain user names
and paths as well. See details below.}
}
\value{
Character vector of environment variables.
}
\description{
gitcreds caches credentials in environment variables.
}
\details{

}
\seealso{
\code{\link[=gitcreds_get]{gitcreds_get()}}.
}
