% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-auth.R
\name{gitcreds_get}
\alias{gitcreds_get}
\title{Query git credentials}
\usage{
gitcreds_get(url = "https://github.com")
}
\arguments{
\item{url}{URL to get credentials for.}
}
\description{
Query git credentials
}
